requirements.txt
fastapi==0.59.0                 # Framework web
uvicorn==0.11.5                 # Servidor
psycopg2-binary~=2.8.6